src/backend/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "Backend to stock sentiment tracker platform"
authors = ["Harshaa Vardaan"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
python-dotenv = "^0.20.0"
tweepy = {extras = ["async"], version = "^4.10.0"}
praw = "^7.6.0"
transformers = "^4.21.2"
emoji = "^2.1.0"
celery = "^5.2.7"
influxdb-client = "^1.33.0"
nltk = "^3.7"
scipy = "^1.9.2"
yfinance = "^0.1.85"
fastapi-redis-cache = "^0.2.5"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.16.4"
ipython = "^8.5.0"
flower = "^1.2.0"
black = "^22.10.0"
pylint = "^2.15.5"


[tool.poe.tasks]
download-pytorch-cuda11 = "pip install torch==1.12.1+cu116 --extra-index-url https://download.pytorch.org/whl/cu116"
download-pytorch-cpu = "pip install torch==1.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"